Converting feed detail actionsheet to popover. Fixing all ios 6 warnings.

+++ media/ios/ShareKit/Core/Helpers/OAuth/OAMutableURLRequest.m
@@ -192,7 +192,7 @@ signatureProvider:(id<OASignatureProviding, NSObject>)aProvider
 
 - (void)_generateTimestamp
 {
-    timestamp = [[NSString stringWithFormat:@"%d", time(NULL)] retain];
+    timestamp = [[NSString stringWithFormat:@"%d", (int)time(NULL)] retain];
 }
 
 - (void)_generateNonce
